FV6VS2M split git.test.ts by concern so vitest can spread it
